refactor(core): Split agent execution orchestrator into chat/task and workflow services (no-changelog)
